requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
cachetools==4.2.2
certifi==2021.10.8
Django==3.2.3
django-environ==0.8.1
django-modeltranslation==0.17.5
gunicorn==20.1.0
Pillow==9.0.0
psycopg2-binary
python-telegram-bot==13.10
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
tzlocal==4.1
